+++ server/routes/index.js
@@ -7,7 +7,7 @@ module.exports = app => {
   app.get(`${config.apiBase}`, (req, res, next) => {
     res.send({
       hydrusrv: {
-        version: '2.7.1'
+        version: '2.8.0'
       }
     })
   })