+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [2.8.0] - 2018-07-28
+
 ### Changed
 
 + Refactored files querying
@@ -254,7 +256,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 + Initial release
 
-[Unreleased]: https://github.com/mserajnik/hydrusrv/compare/2.7.1...develop
+[Unreleased]: https://github.com/mserajnik/hydrusrv/compare/2.8.0...develop
+[2.8.0]: https://github.com/mserajnik/hydrusrv/compare/2.7.1...2.8.0
 [2.7.1]: https://github.com/mserajnik/hydrusrv/compare/2.7.0...2.7.1
 [2.7.0]: https://github.com/mserajnik/hydrusrv/compare/2.6.1...2.7.0
 [2.6.1]: https://github.com/mserajnik/hydrusrv/compare/2.6.0...2.6.1